+++ src/tasks/admin/index.js
@@ -15,6 +15,7 @@ const initListUsers = require('./list-users');
 /**
  * @param {AuthTokenRules} authTokenRules The auth token rules
  * @param {RealmRepository} realmRepository The realm repository
+ * @param {RealmLimitsRepository} realmLimitsRepository The realm limits repository
  * @param {AuthRepository} authRepository The auth repository
  * @param {ChannelRepository} channelRepository The channel repository
  * @param {RealtimeConnectionRepository} realtimeConnectionRepository The real-time connection repository
@@ -28,6 +29,7 @@ module.exports = ({
   authRepository,
   channelRepository,
   realmRepository,
+  realmLimitsRepository,
   realtimeConnectionRepository,
   subscriptionRepository,
   userRepository,
@@ -35,7 +37,7 @@ module.exports = ({
 }) => ({
   fetchRealm: initFetchRealm({realmRepository}),
   createChannel: initCreateChannel({realmRepository, channelRepository}),
-  createRealm: initCreateRealm({realmRepository}),
+  createRealm: initCreateRealm({realmRepository, realmLimitsRepository}),
   createRealmToken: initCreateRealmToken({authTokenRules, realmRepository, userRepository, authRepository}),
   createSubscription: initCreateSubscription({
     userRepository,